+++ src/OneSignalSDKWorker.js
@@ -1,4 +1,4 @@
 // IMPORT PROJECT SERVICE WORKER
-importScripts('service-worker.js');
+importScripts('ngsw-worker.js');
 // IMPORT OneSignal FOR PUSH NOTIFICATIONS
 importScripts('https://cdn.onesignal.com/sdks/OneSignalSDK.js');