+++ src/service-worker.js
@@ -1,1 +1,2 @@
 console.log('in was-tutorial service worker');
+importScripts('OneSignalSDKWorker.js');